// src/app/store/content/types/employees/EmployeesActionTypes.enum.ts
export enum EEmployeesActionTypes {
  FIND_ALL = '[Employee]-find-all',
  FIND_BY_ID = '[Employee]-find-by-id',
  CREATE = '[Employee]-create',
  DELETE = '[Employee]-delete',
  UPDATE = '[Employee]-update',
  SUCCESS_GET_ONE = '[Employee]-get-one-success',
  SUCCESS_GET_MANY = '[Employee]-get-many-success',
  SUCCESS_DELETE = '[Employee]-delete-success',
  ERROR = '[Employee]-failure',
}